Custom Scripts added to create Fake Customers and Products

+++ app/store/models.py
@@ -13,7 +13,7 @@ class Customer(models.Model):
 
 class Product(models.Model):
     name = models.CharField(max_length=200,null=False,unique=True)
-    price = models.PositiveIntegerField(null=False)
+    price = models.FloatField(null=False)
 
     def __str__(self):
         return self.name
